+++ packages/metastable/src/sysinfo/disk.ts
@@ -3,8 +3,8 @@ import { statfs } from 'fs/promises';
 export async function usage(path: string) {
   const stat = await statfs(path);
   return {
-    diskPath: '',
-    size: stat.blocks * stat.bsize,
+    path,
+    total: stat.blocks * stat.bsize,
     free: stat.bfree * stat.bsize,
     used: (stat.blocks - stat.bfree) * stat.bsize,
   };