+++ packages/types/src/instance.ts
@@ -23,6 +23,7 @@ export interface InstanceInfo {
   vram: number;
   dataRoot: string;
   features: Feature[];
+  defaultDirectory: string;
 }
 
 export interface UpdateInfo {